Add build and run scripts generator
The `generate.py` script will be downloaded from
https://raw.githubusercontent.com/unikraft/app-testing/staging/scripts/generate.py -O scripts/generate.py
and will be used to generate the scripts in the `scripts/build/` and `scripts/run/`
directories. This will improve maintenance (only required to happen for the
`generate.py` script), while still keeping things usable by providing the basic
build and run scripts.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>